+++ examples/groq/search_graph_groq_openai.py
@@ -4,7 +4,7 @@ Basic example of scraping pipeline using SmartScraper
 
 import os
 from dotenv import load_dotenv
-from scrapegraphai.graphs import SmartScraperGraph
+from scrapegraphai.graphs import SearchGraph
 from scrapegraphai.utils import prettify_exec_info
 
 load_dotenv()
@@ -14,33 +14,33 @@ load_dotenv()
 # Define the configuration for the graph
 # ************************************************
 
+groq_key = os.getenv("GROQ_APIKEY")
 openai_key = os.getenv("OPENAI_APIKEY")
 
 graph_config = {
     "llm": {
-        "api_key": openai_key,
-        "model": "gpt-3.5-turbo",
+        "model": "groq/gemma-7b-it",
+        "api_key": groq_key,
+        "temperature": 0
     },
-    "verbose": True,
+    "embeddings": {
+        "api_key": openai_key,
+        "model": "openai",
+    },
+    "headless": False
 }
 
-# ************************************************
-# Create the SmartScraperGraph instance and run it
-# ************************************************
-
-smart_scraper_graph = SmartScraperGraph(
-    prompt="List me all the projects with their description.",
-    # also accepts a string with the already downloaded HTML code
-    source="https://perinim.github.io/projects/",
+search_graph = SearchGraph(
+    prompt="List me the best escursions near Trento",
     config=graph_config
 )
 
-result = smart_scraper_graph.run()
+result = search_graph.run()
 print(result)
 
 # ************************************************
 # Get graph execution info
 # ************************************************
 
-graph_exec_info = smart_scraper_graph.get_execution_info()
+graph_exec_info = search_graph.get_execution_info()
 print(prettify_exec_info(graph_exec_info))