+++ examples/local_models/json_scraper_ollama.py
@@ -1,18 +1,18 @@
 """
-Basic example of scraping pipeline using XMLScraperGraph from XML documents
+Basic example of scraping pipeline using JSONScraperGraph from JSON documents
 """
 
 import os
 from dotenv import load_dotenv
-from scrapegraphai.graphs import XMLScraperGraph
+from scrapegraphai.graphs import JSONScraperGraph
 from scrapegraphai.utils import convert_to_csv, convert_to_json, prettify_exec_info
 load_dotenv()
 
 # ************************************************
-# Read the XML file
+# Read the JSON file
 # ************************************************
 
-FILE_NAME = "inputs/books.xml"
+FILE_NAME = "inputs/example.json"
 curr_dir = os.path.dirname(os.path.realpath(__file__))
 file_path = os.path.join(curr_dir, FILE_NAME)
 
@@ -35,27 +35,28 @@ graph_config = {
         "model": "ollama/nomic-embed-text",
         "temperature": 0,
         "base_url": "http://localhost:11434",
-    }
+    },
+    "verbose": True,
 }
 
 # ************************************************
-# Create the XMLScraperGraph instance and run it
+# Create the JSONScraperGraph instance and run it
 # ************************************************
 
-xml_scraper_graph = XMLScraperGraph(
+json_scraper_graph = JSONScraperGraph(
     prompt="List me all the authors, title and genres of the books",
     source=text,  # Pass the content of the file, not the file object
     config=graph_config
 )
 
-result = xml_scraper_graph.run()
+result = json_scraper_graph.run()
 print(result)
 
 # ************************************************
 # Get graph execution info
 # ************************************************
 
-graph_exec_info = xml_scraper_graph.get_execution_info()
+graph_exec_info = json_scraper_graph.get_execution_info()
 print(prettify_exec_info(graph_exec_info))
 
 # Save to json or csv