+++ examples/local_models/scrape_plain_text_ollama.py
@@ -1,21 +1,20 @@
-"""
-Basic example of scraping pipeline using JSONScraperGraph from JSON documents
+""" 
+Basic example of scraping pipeline using SmartScraper from text
 """
 
 import os
-from dotenv import load_dotenv
-from scrapegraphai.graphs import JSONScraperGraph
-from scrapegraphai.utils import convert_to_csv, convert_to_json, prettify_exec_info
-load_dotenv()
+from scrapegraphai.graphs import SmartScraperGraph
+from scrapegraphai.utils import prettify_exec_info
 
 # ************************************************
-# Read the JSON file
+# Read the text file
 # ************************************************
 
-FILE_NAME = "inputs/example.json"
+FILE_NAME = "inputs/plain_html_example.txt"
 curr_dir = os.path.dirname(os.path.realpath(__file__))
 file_path = os.path.join(curr_dir, FILE_NAME)
 
+# It could be also a http request using the request model
 with open(file_path, 'r', encoding="utf-8") as file:
     text = file.read()
 
@@ -29,33 +28,32 @@ graph_config = {
         "temperature": 0,
         "format": "json",  # Ollama needs the format to be specified explicitly
         # "model_tokens": 2000, # set context length arbitrarily
+        "base_url": "http://localhost:11434",
     },
     "embeddings": {
         "model": "ollama/nomic-embed-text",
         "temperature": 0,
-    }
+        "base_url": "http://localhost:11434",
+    },
+    "verbose": True,
 }
 
 # ************************************************
-# Create the JSONScraperGraph instance and run it
+# Create the SmartScraperGraph instance and run it
 # ************************************************
 
-json_scraper_graph = JSONScraperGraph(
-    prompt="List me all the authors, title and genres of the books",
-    source=text,  # Pass the content of the file, not the file object
+smart_scraper_graph = SmartScraperGraph(
+    prompt="List me all the projects",
+    source=text,
     config=graph_config
 )
 
-result = json_scraper_graph.run()
+result = smart_scraper_graph.run()
 print(result)
 
 # ************************************************
 # Get graph execution info
 # ************************************************
 
-graph_exec_info = json_scraper_graph.get_execution_info()
+graph_exec_info = smart_scraper_graph.get_execution_info()
 print(prettify_exec_info(graph_exec_info))
-
-# Save to json or csv
-convert_to_csv(result, "result")
-convert_to_json(result, "result")