+++ examples/local_models/scrape_xml_ollama.py
@@ -1,12 +1,9 @@
 """
-Basic example of scraping pipeline using XMLScraperGraph from XML documents
+Basic example of scraping pipeline using SmartScraper from XML documents
 """
-
 import os
-from dotenv import load_dotenv
-from scrapegraphai.graphs import XMLScraperGraph
-from scrapegraphai.utils import convert_to_csv, convert_to_json, prettify_exec_info
-load_dotenv()
+from scrapegraphai.graphs import SmartScraperGraph
+from scrapegraphai.utils import prettify_exec_info
 
 # ************************************************
 # Read the XML file
@@ -23,39 +20,40 @@ with open(file_path, 'r', encoding="utf-8") as file:
 # Define the configuration for the graph
 # ************************************************
 
+
 graph_config = {
     "llm": {
         "model": "ollama/mistral",
         "temperature": 0,
         "format": "json",  # Ollama needs the format to be specified explicitly
         # "model_tokens": 2000, # set context length arbitrarily
+        "base_url": "http://localhost:11434",  # set ollama URL arbitrarily
     },
     "embeddings": {
         "model": "ollama/nomic-embed-text",
         "temperature": 0,
-    }
+        "base_url": "http://localhost:11434",  # set ollama URL arbitrarily
+    },
+    "verbose": True,
 }
 
 # ************************************************
-# Create the XMLScraperGraph instance and run it
+# Create the SmartScraperGraph instance and run it
 # ************************************************
 
-xml_scraper_graph = XMLScraperGraph(
+smart_scraper_graph = SmartScraperGraph(
     prompt="List me all the authors, title and genres of the books",
     source=text,  # Pass the content of the file, not the file object
     config=graph_config
 )
 
-result = xml_scraper_graph.run()
+result = smart_scraper_graph.run()
 print(result)
+
 
 # ************************************************
 # Get graph execution info
 # ************************************************
 
-graph_exec_info = xml_scraper_graph.get_execution_info()
+graph_exec_info = smart_scraper_graph.get_execution_info()
 print(prettify_exec_info(graph_exec_info))
-
-# Save to json or csv
-convert_to_csv(result, "result")
-convert_to_json(result, "result")