+++ scrapegraphai/graphs/__init__.py
@@ -2,6 +2,7 @@
 __init__.py file for graphs folder
 """
 
+from .abstract_graph import AbstractGraph
 from .base_graph import BaseGraph
 from .smart_scraper_graph import SmartScraperGraph
 from .speech_graph import SpeechGraph
@@ -10,3 +11,4 @@ from .script_creator_graph import ScriptCreatorGraph
 from .xml_scraper_graph import XMLScraperGraph
 from .json_scraper_graph import JSONScraperGraph
 from .csv_scraper_graph import CSVScraperGraph
+from .pdf_scraper_graph import PDFScraperGraph